+++ packages/perseus/src/widget-ai-utils/interaction/interaction-ai-utils.test.ts
@@ -1,5 +1,7 @@
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./interaction-ai-utils";
+
 import type {PerseusRenderer} from "../../perseus-types";
 
 const question1: PerseusRenderer = {
@@ -262,7 +264,17 @@ const question1: PerseusRenderer = {
     },
 };
 
-describe("interaction widget", () => {
+describe("Interaction AI utils", () => {
+    it("it returns JSON with the expected format and fields", () => {
+        const resultJSON = getPromptJSON();
+
+        expect(resultJSON).toEqual({
+            type: "interaction",
+            isSupported: false,
+            message: "",
+        });
+    });
+
     it("should get prompt json which matches the state of the UI", async () => {
         // Arrange
         const {renderer} = renderQuestion(question1);
